Update TextEncodingHelper.cs
Some of my songs are throwing an error on the method GetString because the method GetSplitStrings is calling it with an empty string array. I tried opening the same files on another ID3 programs and the files open just fine. While not sure why this is happening, preventing the method from processing the empty string mitigates the problem.